refactor(agent-engine): isolate agent interaction runtime inputs
